+++ src/main/java/com/ltk/springproject/repository/ReplyRepository.java
@@ -6,16 +6,16 @@ import com.ltk.springproject.domain.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+public interface ReplyRepository extends JpaRepository<Reply, Long> { // Long으로 변경
     // 특정 게시글에 달린 댓글 목록 찾기 (부모 댓글이 없는 최상위 댓글들)
     List<Reply> findByArticleAndParentIsNull(Article article);
-    List<Reply> findByArticleIdAndParentIsNull(Integer articleId);
+    List<Reply> findByArticleIdAndParentIsNull(Long articleId); // Long으로 변경
 
     // 특정 회원이 작성한 댓글 목록 찾기
     List<Reply> findByMember(Member member);
-    List<Reply> findByMemberId(Long memberId);
+    List<Reply> findByMemberId(Long memberId); // Long으로 변경
 
     // 특정 부모 댓글에 속한 자식 댓글(대댓글) 목록 찾기
     List<Reply> findByParent(Reply parent);
-    List<Reply> findByParentId(Integer parentId);
+    List<Reply> findByParentId(Long parentId); // Long으로 변경
 }
