+++ src/main/java/com/ltk/springproject/repository/WorkRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 // import java.time.LocalDate; // 필요시
 
-public interface WorkRepository extends JpaRepository<Work, Integer> {
+public interface WorkRepository extends JpaRepository<Work, Long> { // Long으로 변경
     // 특정 시리즈에 속한 작품들 찾기
     List<Work> findBySeries(Series series);
 
     // 특정 시리즈 ID로 작품들 찾기
-    List<Work> findBySeriesId(Integer seriesId);
+    List<Work> findBySeriesId(Long seriesId); // Long으로 변경
 
     // 작품 타입으로 작품들 찾기
     List<Work> findByType(String type);
